add dynamic and glsk
