+++ pypowsybl/report/impl/reporter.py
@@ -3,8 +3,9 @@
 # This Source Code Form is subject to the terms of the Mozilla Public
 # License, v. 2.0. If a copy of the MPL was not distributed with this
 # file, You can obtain one at http://mozilla.org/MPL/2.0/.
+# SPDX-License-Identifier: MPL-2.0
 #
-import pypowsybl._pypowsybl as _pp
+import pypowsybl._pypowsybl as _pp # pylint: disable=protected-access
 
 
 class Reporter:  # pylint: disable=too-few-public-methods
