+++ pypowsybl/security/impl/security_analysis_result.py
@@ -11,7 +11,7 @@ from prettytable import PrettyTable
 
 from pypowsybl import _pypowsybl
 from pypowsybl._pypowsybl import PreContingencyResult, PostContingencyResult
-from pypowsybl.util import _create_data_frame_from_series_array
+from pypowsybl.utils.util import _create_data_frame_from_series_array
 
 
 class SecurityAnalysisResult:
